Misc changes

+++ dice.c
@@ -748,7 +748,7 @@ CloseRNG( const rng rngx, void *p ) {
 
 
 extern int
-RNGSystemSeed( const rng rngx, void *p, int *pnSeed ) {
+RNGSystemSeed( const rng rngx, void *p, unsigned long *pnSeed ) {
 
   int h;
   int f = FALSE;
@@ -799,11 +799,12 @@ RNGSystemSeed( const rng rngx, void *p, int *pnSeed ) {
 
 }
 
-extern void *InitRNG( int *pnSeed, int *pfInitFrom,
+extern void *InitRNG( unsigned long *pnSeed, int *pfInitFrom,
                       const int fSet, const rng rngx ) {
 
     int f = FALSE;
     rngcontext *rngctx =  g_new0(rngcontext, 1);
+
     /* misc. initialisation */
 
     /* Mersenne-Twister */
